+++ CHANGES.md
@@ -7,6 +7,8 @@
 * Compiler: fix location for parsing errors when last token is a virtual semicolon
 * Compiler: fix variable renaming with nested const/let decl with identical names
 * Compiler: fix variable renaming inside js method
+* Compiler: consise body should allow any expression but object literals
+* Compiler: preverve [new] without arguments [new C] (vs [new C()]
 
 # 5.2.0 (2023-04-28) - Lille
 ## Features/Changes